+++ TODO.md
@@ -1,10 +1,11 @@
 ## make a back button
-
-## github only sends like 20 repos at a time, 
-## need to make a way to get all of them 
-## or an options to get more/search/filter
 -- 
-
+## github only sends like 20 repos at a time, 
+ - need to make a way to get all of them 
+ - or an options to get more/search/filter
+-- 
+## should add ssh_url to the options and prompt the user to choose 
+ - between https and ssh
 
 
 get cfg
