+++ src/ConfigManager.ts
