+++ src/auth/authenticate.ts
@@ -1,11 +1,11 @@
 import { spawn } from "child_process";
-import { ConfigManager, type UserConfig } from "./config";
-import { authorizationWithGithub } from "./server";
+import { ConfigManager, type UserConfig } from "../ConfigManager";
+import { authorize} from '.'
 
-export async function Authenticate(config: UserConfig, cm: ConfigManager): Promise<void> {
+export async function authenticate(config: UserConfig, cm: ConfigManager): Promise<void> {
   if (config.user_profile.access_token === "") {
     const serverProcess = spawn('node', ['src/server.ts']);
-    await authorizationWithGithub(cm)
+    await authorize(cm)
     serverProcess.on('exit', (code, signal) => (code && signal))
     await cm.revalidateConfig()
   } else {
