+++ src/auth/index.ts
@@ -1,0 +1,4 @@
+
+
+export { authenticate } from './authenticate'
+export { authorizationWithGithub as authorize } from './authorizationWithGithub'
