+++ src/cli/configureEditor/alias_options.ts
