+++ src/index.ts
@@ -1,12 +1,12 @@
-import { ConfigManager } from "./config";
-import { Authenticate } from "./authenticate";
+import { ConfigManager } from "./ConfigManager";
+import { authenticate } from "./auth";
 import { initCli } from './cli'
 
 (async function main() {
   const cm = new ConfigManager()
   const config = cm.getConfig()
 
-  await Authenticate(config, cm)
+  await authenticate(config, cm)
 
   await initCli(config, cm)
 })()
