+++ src/utils/exit.ts
