+++ src/utils/index.ts
@@ -1,0 +1,4 @@
+
+export { Exit as exit } from './exit'
+export { openBrowser } from './openBrowser'
+export { rando } from './rando'
