+++ src/utils/openBrowser.ts
